# api-docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Audio transcription service using WhisperX
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Scriberr API
  version: "1.0"
paths:
...
  /api/v1/transcription/{id}/runs/{run_id}/recovery:
    get:
      description: Return saved stages, attempts, partial text and same-plan resume
        availability
      parameters:
      - description: Recording ID
        in: path
        name: id
        required: true
        type: string
      - description: Execution ID
        in: path
        name: run_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get execution recovery details
      tags:
      - transcription
  /api/v1/transcription/{id}/runs/{run_id}/resume:
    post:
      description: Resume the same execution within its original deadline and attempt
        budget; accepts no quality overrides
      parameters:
      - description: Recording ID
        in: path
        name: id
        required: true
        type: string
      - description: Execution ID
        in: path
        name: run_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Resume a saved execution
      tags:
      - transcription
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: JWT token with Bearer prefix
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"